+++ js/comment.js
@@ -1,0 +1,8 @@
+//здесь будет функционал по добавлению комментов на сайте
+
+function checkCommentLength(comment, maxLength) {
+  return comment.length <= maxLength;
+}
+
+checkCommentLength('Comment test', 140);
+
