+++ js/photo-edit.js
@@ -1,0 +1,1 @@
+//здесь будет функционал по редактировнию фото
