+++ js/photo-input.js
@@ -1,0 +1,1 @@
+//здесь будет функционал по загрузке фото на сайт
