+++ js/photo-view.js
@@ -1,0 +1,1 @@
+//здесь будет функционал по просмотру фото (с фильтрацией)
